First attempt at building archive unzipping
Compiles, but will fail occasionally.
Utilises a variety of wxWidgets provided files to iterate through the
files in an archive, open input streams on the file data in the
archive, and write the contents of the input streams to an output
stream pointed to a new file in the target location.

This appears to fail randomly, however testing hasn't been very
comprehensive. Primary failures appear to be an inability to open
output file streams to the target file - presumably due to a
permissions or incomplete path.