Fixed issue adding certain mod folders
The issues lay in the fact that whilst the #getAllMods 'cleaned' any mod config JSON files that it attempted to parse, the #onAddModFolder handler did not do so, and would encounter cases of JSON standards violations (comments and trailing commas) causing the parser to throw an exception.

The fix presented is to modify the mmConfig class to provide a gneralised interface for interacting with JSON files in the application's specific scenario - thus, consistently cleaning any JSON files encountered, and improving code reuse.

This commit also includes a fix for cases where the mods directory does not contain an "enabled_mods.json" file, which would cause the application to fail. Instead, if the enabled mods file does not exist upon the installation of a mod, a new one is generateted instead.